Common: add common files for all the platforms.
This commit is squashed and merged from branches
wmo/v2014.06.0 for Basis/Corazon/Coronet support and
wmo/2016.05.work for Dagger support.
The git history is retained in those branches.

+++ commands/memtest.c
@@ -35,7 +35,7 @@ static int do_test_one_area(struct mem_test_resource *r, int bus_only,
 
 	printf("Testing memory space: "
 			"0x%08x -> 0x%08x:\n",
-			r->r->start,  r->r->end);
+			(size_t)r->r->start,  (size_t)r->r->end);
 
 	remap_range((void *)r->r->start, r->r->end -
 			r->r->start + 1, cache_flag);
